Dedupe MoleDescriptor/Options type
