+++ src/platform-implementation-js/namespaces/widgets.ts
@@ -8,7 +8,7 @@ import ComposeView from '../views/compose-view';
 import type { Driver, DrawerViewOptions } from '../driver-interfaces/driver';
 import get from '../../common/get-or-fail';
 import isComposeTitleBarLightColor from '../dom-driver/gmail/is-compose-titlebar-light-color';
-import { MoleOptions } from '../driver-interfaces/mole-view-driver';
+import { type MoleOptions } from '../dom-driver/gmail/widgets/gmail-mole-view-driver';
 import type { Widgets as IWidgets, ModalDescriptor } from '../../inboxsdk';
 
 class Widgets implements IWidgets {
